extract booklist into component
