+++ src/app/screens/HomeScreen/index.tsx
@@ -1,19 +1,12 @@
 import React, { useEffect } from 'react';
-import { FlatList, ListRenderItem } from 'react-native';
 import { useDispatch, useSelector } from 'react-redux';
-import BookCell from '@components/BookCell';
-import { Book } from '@interfaces/book';
-import { Navigation } from '@interfaces/navigation';
-import { DETAIL_SCREEN } from '@constants/routes';
-import { NAV_MARGIN } from '@constants/commonStyles';
 import { AppState } from '@interfaces/redux';
 import actionCreators from '@redux/books/actions';
 import WithLoader from '@components/WithLoader';
 import WithErrorView from '@components/WithErrorView';
+import BookList from '@app/components/BookList';
 
-import styles from './styles';
-
-function HomeScreen({ navigation }: Navigation<any>) {
+function HomeScreen() {
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(actionCreators.getBooks());
@@ -21,19 +14,10 @@ function HomeScreen({ navigation }: Navigation<any>) {
   const books = useSelector((state: AppState) => state.book.books);
   const loading = useSelector((state: AppState) => state.book.booksLoading);
   const error = useSelector((state: AppState) => state.book.booksError);
-  const onBookPress: (book: Book) => void = book => navigation.navigate(DETAIL_SCREEN, book);
-  const renderBook: ListRenderItem<Book> = ({ item }) => <BookCell book={item} onBookPress={onBookPress} />;
-  const keyExtractor = ({ id }: Book) => String(id);
   return (
     <WithLoader loading={loading}>
       <WithErrorView error={error}>
-        <FlatList<Book>
-          style={styles.scrollView}
-          data={books}
-          renderItem={renderBook}
-          keyExtractor={keyExtractor}
-          contentContainerStyle={NAV_MARGIN}
-        />
+        <BookList books={books} />
       </WithErrorView>
     </WithLoader>
   );
