--- src/app/screens/HomeScreen/styles.ts
@@ -1,7 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-  scrollView: {
-    paddingBottom: 5
-  }
-});
